% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after_degree.R
\name{after_degree}
\alias{after_degree}
\title{Gives insight in enrolments year after degree}
\usage{
after_degree(program_of_interest, my_flow_insights, prop_exam = FALSE)
}
\arguments{
\item{program_of_interest}{string containing the 5-character program code of interest}

\item{my_flow_insights}{a flow_insights object}

\item{prop_exam}{boolean, default FALSE, set to TRUE for after propedeutic exam}
}
\value{
a list with 10 objects
\item{program_of_interest}{the program code that was entered}
\item{degree}{D for propedeutic, A for AD, B for Bachelor or M for Master}
\item{brin_own}{brin_own from the my_flow_insights object}
\item{academic_year_from}{the earliest academic year in the data}
\item{academic_year_to}{the latest academic year in the data}
\item{n_students}{number of students with a degree from brin_own/program_of_interest in academic_year_from}
\item{n_students_enrolment_type_single}{number of above students with at most one enrolment brin/program per academic year}
\item{n_students_enrolment_type_multiple}{difference between the two numbers of students}
\item{from_to_single}{dataframe with 7 columns with counts per flow, for the students with enrolment_type = single}
\item{from_to_multiple}{dataframe with 7 columns with flow information per (anonymized) student, for the students with enrolment_type = single }
}
\description{
Gives insight in enrolments year after degree
}
\examples{
\dontrun{
after_degree(program_of_interest = "34507",
 my_flow_insights = my_insights_object,
  prop_exam = TRUE)
}
}
