% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bek.R
\name{bek}
\alias{bek}
\title{Makes a bek object}
\usage{
bek(df = data.frame())
}
\arguments{
\item{df}{a data.frame resulting from using the read_bek_data() on the funding csv file}
}
\value{
a list with 6 objects:
\item{brin_own}{text containing the brin of the higher educational institution to which the funding file refers to}
\item{year_funding}{text containing the year the funding file refers to}
\item{date_retrieval}{date containing the date the funding file was retrieved from DUO}
\item{students}{data.frame containing id's from persons with a least one enrolment or degree from the brin concerned }
\item{enrolments}{data.frame containing data of enrolments of these students}
\item{degrees}{data.frame containing data of acquired degrees of these students }
}
\description{
Takes the data.frame resulting from read_bek_data() and returns a list object
}
\examples{
\dontrun{
bek(my_bek_data)
}
}
