% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact_bek.R
\name{compact_bek}
\alias{compact_bek}
\title{Makes a compacter version of a bek object, with some summaries}
\usage{
compact_bek(my_bek)
}
\arguments{
\item{my_bek}{a bek object}
}
\value{
a bek_compact_object as a list with 8 objects
\item{type}{text containing the type of bek file the data is from}
\item{brin_own}{text containing the brin of the higher educational institution to which the funding file refers to}
\item{enrolments_degrees_compact}{data frame with one row per academic year, student, brin and program_code, adorned with date_degree when applicable and a note if the student has at most a single enrolment per year}
\item{presences_brin}{dataframe with one row per student per academic year displaying enrolmentinformation regarding brin: either in brin_own and/or other HE, or outside HE }
\item{presences_level}{dataframe with one row per student per academic year displaying enrolmentinformation regarding level: HBO-AD, HBO-BA, HBO-MA, WO-BA, WO-MA }
\item{switches}{dataframe with columns from_brin, from_program_code, to_brin_to_program_code, n_students for all students in df enrolment_degrees_compact, that have at most one enrolment per year, and did not receive a diploma in first year}
\item{summary_presences_brin}{presences_brin per academic year summarised over students}
\item{summary_presences_level}{presences_leveln per academic year summarised over students}
}
\description{
Makes a compacter version of a bek object, with some summaries
}
\examples{
\dontrun{
bek_compact(my_bek)
}
}
