% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_flows.R
\name{find_flows}
\alias{find_flows}
\title{Extracts switches or stacks between own brin and another brin}
\usage{
find_flows(
  my_flow_insights,
  stack_or_switch = c("switch", "stack"),
  direction = c("to", "from"),
  other_brin
)
}
\arguments{
\item{my_flow_insights}{a flow_insights object}

\item{stack_or_switch}{either "switch" for flows without final degree or "stack" for flows  with final degree}

\item{direction}{"to" the other_brin for outgoing flows, or "from" the other brin for incoming flows}

\item{other_brin}{the brin from another institution of higher education}
}
\value{
a dataframe with a subset of the switches or stacks dataframe from the flow_insights object
}
\description{
Extracts switches or stacks between own brin and another brin
}
\examples{
\dontrun{
find_flows(my_insights_object, "switch", "to", "98XX")
}
}
