% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_metrics.R
\name{key_metrics}
\alias{key_metrics}
\title{Calculates summary metrics from vlpbek object}
\usage{
key_metrics(my_vlpbek, label_brin_own = "no", language_choice = "EN")
}
\arguments{
\item{my_vlpbek}{a vlpbek object}

\item{label_brin_own}{an optional text to use instead of the own BRIN,}

\item{language_choice}{is either "EN" (default) or "NL",}
}
\value{
a list with 12 objects
\item{brin_own}{text containing the brin of the higher educational institution to which the funding file refers to}
\item{year_funding}{text containing the year the funding file refers to}
\item{date_retrieval}{date containing the date the funding file was retrieved from DUO}
\item{academic_years}{vector with the two academic years mentioned in enrolments }
\item{n_student}{number of different students mentioned in enrolments}
\item{from_BRIN}{data.frame containing BRINs where students were enrolled in the older academic year and count of those students}
\item{to_BRIN}{data.frame containing BRINs where students were enrolled in the newer academic year and count of those students}
\item{n_program_code}{number of different program codes mentioned in enrolments}
\item{n_BRIN_program_code}{number of different combination BRIN/ program code mentioned in enrolments}
\item{n_student_BRIN_program_code}{number of different combination student / BRIN/ program code mentioned in enrolments}
\item{brin_flows}{data.frame containing enrolment flows from older academic year to newer academic year with regard to being enrolled in own brin or other HEI}
}
\description{
Calculates summary metrics from vlpbek object
}
\examples{
\dontrun{
key_metrics(my_vlpbek, label_brin_own = "My institution")
}
}
