% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankeygradient.R
\name{plot_sankeygradient}
\alias{plot_sankeygradient}
\title{Plots a sankey diagram with gradient colors, based on geom_ribbon}
\usage{
plot_sankeygradient(
  column_from,
  column_to,
  color_from,
  color_to,
  values,
  padding = 2,
  alpha = 0.4,
  display_labels = TRUE,
  match_label_color = TRUE,
  label_fontface = "bold",
  label_size = 10
)
}
\arguments{
\item{column_from}{vector of character labels, indicating the lefthand side}

\item{column_to}{vector of character labels, indicating the righthand side}

\item{color_from}{vector of color values, indicating colors for the lefthand side}

\item{color_to}{vector of color values, indicating colors for the righthand side}

\item{values}{vector of values, indicating the width of the ribbon}

\item{padding}{constant used for displaying space between ribbons}

\item{alpha}{numeric, transparancy of colors (0 = complete transparent, 1 = opaque)}

\item{display_labels}{boolean used for displaying labels or not}

\item{match_label_color}{boolean used to indicate labels in same color as nodes}

\item{label_fontface}{string used for styling the font of labels}

\item{label_size}{number indicating the font size of labels}
}
\value{
a ggplot with geom_ribbons
}
\description{
Plots a sankey diagram with gradient colors, based on geom_ribbon
}
\examples{
\dontrun{
plot_sankeygradient(column_from = c("a","b"),
                   column_to = c("d","d"),
                   values = c(5,10),
                   color_from = c("red", "green"),
                   color_to = c("blue", "blue"))
}
}
