% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bek_data.R
\name{read_bek_data}
\alias{read_bek_data}
\title{Read in a bek file}
\usage{
read_bek_data(locationname_bek)
}
\arguments{
\item{locationname_bek}{location of the bek type of file (bekostigingsbestand)}
}
\value{
a dataframe containing the raw bek file, with attribute regarding type (vlp, def, his)
}
\description{
\code{read_bek_data} reads in a bek type of file (bekostigingsbestand), containing data about funding.
}
