% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vlpbek_data.R
\name{read_vlpbek_data}
\alias{read_vlpbek_data}
\title{Read in a vlpbek file}
\usage{
read_vlpbek_data(locationname_vlpbek)
}
\arguments{
\item{locationname_vlpbek}{location of the vlpbek type of file (bekostigingsbestand)}
}
\value{
a dataframe containing the raw vlpbek file
}
\description{
\code{read_vlpbek_data} reads in a vlpbek type of file (bekostigingsbestand), containing interim data about funding.
The function opens a window with which a file can be chosen.
}
