% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vlpbek.R
\name{read_vlpbek}
\alias{read_vlpbek}
\title{Read in a vlpbek file}
\usage{
read_vlpbek()
}
\value{
a dataframe containing the raw vlpbek file
}
\description{
\code{read_vlpbek} reads in a vlpbek type of file (bekostigingsbestand), containing interim data about funding.
The function opens a window with which a file can be chosen.
}
